New system for defining user-modifiable variables.
Currently I'm only using it for one: pts.  However, the idea is to convert many
of the OpenBuffer and EditorState flags into using this.  This should be very
efficient to check and also lets us expose documentation for each variable.

It doesn't yet persist the values of the variables, and the edition can still be
improved to e.g. do autocomplete on variable names.  But this is already much
better than what was before (just fields directly in the struct, without any
metadata to allow the user to modify them).