Start implementing 'undo' functionality.
For now it only works with delete chars and delete lines and will likely crash
if mixed with other commands.  I expect to finish converting all the other
commands (anything that modifies the contents of the buffer) to the new system
and then it should just work.

This also broke the 'paste' functionality.  It should be restored in a follow
up.