Start switching implementation of editor commands to be in VM.
The hope is that we'll extend map_mode to make it possible to redefine this.  I
want to make it trivial to tweak the commands, so easy that I will actually do
this as part of normal editing. Also, I want to have "recording" mode actually
produce a script (so that one can modify the script itself by changing its
code).

I switched "ar" to this (editor.ReloadCurrentBuffer()) as a proof of concept.