Significantly reworked the search implementation.
Search will now search the entire buffer and only afterwards limit the results
to the desired region; that could be optimized in the future.

It now honors SearchOptions::limit_position. If you select a region with cursors
and enter cursor mode, it will only create cursors in the matches contained
within the region.

I also improved the search predictions buffer somewhat, but it's still
suboptimal (it doesn't yet autocomplete, strangely).