Start adding rudimentary support for a parse tree.
Whenever the buffer contents are updated, we re-parse it. The parsing right now
is terribly simplistic: it creates an entry for each line, then an entry for
each word (honoring the word_characters variable).

"t" enters tree mode. At any point, one is at a given depth of the tree and the
current cursor determines the child in the tree selected. "tj" or "tk" go down
or up in the tree. "th" or "tl" select the previous/next sibling at the current
level (takes repetitions). "td" deletes the entire contents of the current
children (takes repetitions). The terminal displays the current tree (as long as
it's not the entire buffer: the root of the tree).