Change the structure for cursors; pass them explicitly to transformations.
Cursors were represented by a list of pairs of iterators (to the line in the
OpenBuffer::contents_ container) and the column position. Now they are just a
multiset of LineColumn, which is much simpler.

Also adjusted Transformation::Apply to receive explicitly (through the Results
parameter) the cursor to which the transformation should be applied.

I also added an AdjustCursors function, which transformations can use; that
seems much cleaner than having them all go over the underlying structures
directly.