Add completion mode based on the parse tree.
If the dictionary variable is not set and the parse_tree variable is set, \t in
insert mode will attempt to autocomplete from all the words in leaves in the
current parse tree (ignoring leaves that span multiple lines).

While at it, made insert mode not re-parse the tree for every characters (which
was too expensive for large buffers).