Enter (activate link): Search for files also in list of open buffers.
If the path activate doesn't match a file, search in the list of open buffers
for one that matches. The suffix of the path of the buffer must be equal to the
string on which Enter was pressed in order for the match to happen. This also
takes into account positions.

For example, a buffer '/home/alejo/edge/src/line.h' will match the string
'line.h:58:34' (and the cursor will be taken to that position).

The expectation is that this will make it easier to work with languages like
Java or Python where the stack traces rarely include entire paths (often just
the basename).