Drastic refactor for C++ syntax highlighting.
The new implementation is significantly more complex. However, it paves the way
for allowing the customer to control its execution (though that's not yet
exposed through its interface). My goal is to (1) add an interface that allows
the customer to control that (i.e. to parse line-by-line), and (2) make lines
remember recent states, so that we can significantly reduce the amount of time
spent reparsing (i.e. start parsing only at the first line that has been
modified, and if we detect that the current parse information for a line
afterwards is already current, stop parsing further).