Add variable `typos`.
It's currently only moderately useful. For example, in cpp mode, typos inside
comments won't be detected. Inside of double-quoted strings, they will be
detected by the highlighting of the double-quoted string takes precedence.
However, this implements the bulk of the functionality, the other bits should
be improvements to other parts of the system:

- Make comments be parsed recursively (with the words parser).
- Make more-specific highlights take precedence over less-specific highlights.