Offload some of the work of reading contents to background threads.
In particular, once a chunk of bytes have been read and converted into a vector
of wchar_t, the act of breaking into lines and creating the corresponding
`Line` instances to insert is now done in a background thread. The insertion
of the `Line` instances into the buffer is still done in the main thread.