Terminal: Simplify and use HorizontalSplitOutputProducer.
Use HorizontalSplitOutputProducer to combine the output from the buffer tree
(i.e., the main contents) and the editor status. This has the nice advantage of
forcing us to declare where the cursor belongs, which solves a bug when the
cursor jumps back to the buffer tree (because previously we were unaware of the
"accident" that the status had stolen it).

While at it, simplify Terminal::WriteLine which was a bit convoluted.