Navigate mode: support numbers.
They narrow down the range to the corresponding "slice". When a number is
pressed, the active range is divided into 9 slices (with some overlap). The
number brings down the active range to the ith slice.

The slices overlap because I figured it's friendlier (at the cost of slightly
bigger resulting ranges) than requiring the exact slice to be entered. This
allows the user to eye-ball it.