FileDescriptorReader: Add `state_` field.
Use it to avoid reading while lines are being parsed (by a background thread).
This allows me to defer the end-of-file detection until the time when the lines
have been successfully parsed (and inserted into the buffer). Without this, the
call to `read` that finds the end-of-file would happen before those lines have
been inserted, which means that the end-of-file observers would get notified a
bit too soon: after all contents have been successfully read from the file, but
before they've been successfully inserted into the buffer.