Improve the buffer search for `agw`.
Now it tokenizes the names of buffers and confirms that the filter word is a
prefix of any of the tokens.

A name such as `/home/alejo/VolumeConditionJob.java` is tokenized to:

*   `home/alejo/VolumeConditionJob.java`
*   `alejo/VolumeConditionJob.java`
*   `VolumeConditionJob.java`
*   `ConditionJob.java`
*   `Job.java`
*   `java`

(Notice the break of `VolumeConditionJob` into 3 parts.)

So it will be matched by any of the following sets of filter words:

*   `java volu`
*   `job alejo`
*   `conditionjob.java`

It won't be matched by:

*   `ava`
*   `ondition`

(The match has to start at the beginning of one of the tokens.)

This seems more useful than just "the name has to contain the token anywhere"
because I think it's more natural to always start filters from the beginning of
the "logical" words in the path, rather than half-way, so this excludes
accidental matches that are likely to be irrelevant.