Constrain the maximum buffers list lines (at most half of the screen).
This matters when the user opens sooo many buffers that the list of buffers
starts to consume a large portion of the screen. In this case, the list is
constrained and, instead, we start reducing the number of columns that each item
in the list (the name of a buffer) receives.