Add `search_filter_buffer` variable.
If set, searching will delete a buffer (from the set of open buffers, i.e., it
will close it) if it doesn't find any matches. It can be toggled with `v/d`
(variable > search > delete).

This is intended to use with `--view=all`: you can open all files in a
directory, search for a given regular expression, and then edit directly
(knowing that files that didn't have any matches will have been closed).