Improve `:` style expansion: delete the command name before running it.
In insert mode, if the user presses tab after a colon, the previous symbol is
looked up in the VM environment for the buffer and executed. This change it
makes it so that the "previous symbol" (i.e. that identifies the command to
run) is deleted.

This is useful for Zettelkasten command "zkl". The user can now type "0az zkl:";
pressing tab there will delete the " zkl:" part and run the command on "0az".