Support `key` bindings to set buffer_variables of type string.
In the past, the `key` property was only honored for variables of type int or
bool (to toggle them, or set the based on repetitions). For variables of type
string, this just brings up a prompt allowing the user to change their value.

The example is `vn`, to set the namespaces searched.

Make it possible to