Adjust ZK interface to use namespaces.
Now all logic is inside the `zettelkasten` namespace, and function names are
cleaner.

To use it, it makes sense to add it to `cpp_prompt_namespaces`; once you do
that, you no longer need to type the `zk` prefix. I haven't yet figured out the
automation for adding `zettelkasten` to the `cpp_prompt_namespaces`, but that
shouldn't be too hard.

I like that this (1) makes the code more readable (no need for those prefixes),
and also makes it easier to invoke. :-)