Drastically adjust repetitions semantics.
The new semantics should be better at stopping when things overflow. In other
words, if there's a very big number of repetitions followed by a small number in
the opposite direction, the second number will start from wherever the former
number stopped making changes.

This is most evident with the 'f' command: type a number of repetitions that is
significantly larger than the number of occurrences of the character; followed
by a small number.