Remove unnecessary MakeNonNullShared function.
Instead, add a no-arguments constructor for NonNull<std::shared_ptr<>>, which
internally calls std::make_shared.

I'm making the constructor implicit, allowing initialization lists to be used.
This makes things significantly shorter; the thinking is simply that the
NonNull<std::shared<>> types default to a pointer to a newly built instance,
which is pretty useful.