Define gc::WeakPtr.
This should let us keep references that never manage to keep the objects alive.
Ptr would be similar, but it lacks the ability to query whether or not the
object is still alive; adding a separate class explicitly for this feels
safer/simpler: let's us preserve the invariant that if you have a reachable Ptr,
the object is alive (because you made sure to register through some root).