Optimization: Add CleanEden method.
Avoid walking the entire graph of survivors when objects are automatically
collected; in this case, just clean up the lists from the Eden directly.

This is expected to be the norm for objects that are not stored in loops. They
will typically expire directly, without the need for a full-blown GC. Because we
expect this to cover the vast majority of objects, this significantly speeds
things up.

This isn't yet hooked to Edge, this just adds the capability.