Extend vm::NewCallback to handle futures::ValueOrError<T> outputs.
The functions can now return a future; if they do, it must be an
future::ValueOrError<T>, where T has a VMTypeMapper. vm::NewCallback will use
the VMTypeMapper to convert T instances to vm::Value instances (and handle
errors).

I expect this will let me replace a lot of calls to vm::NewFunction with calls
to vm::NewCallback (to allow the types to be deduced, rather than redundantly
given and manually converted).