Avoid wiggling when editor status appears/disappears.
Instead, the editor status now "steals" a line from the output of the lower
widget. In other words, it is a pop-up.

Moved the list of open buffers to be at the top. That means the editor status
won't be stealing from that; otherwise, "ag" would become quite annoying
(because it causes a status to be shown, thus obscuring the view of the open
buffers).

One problem that this introduces is that the editor status now obscures the
current buffer's status. We could/should introduce logic to deal with that in
a better way, but it's not 100% clear yet how to do that. Perhaps we convey to
the buffer that the view is 50 lines, however it really will only get to render
39 lines (maybe cause the context status is also borrowing), and let the render
decide how to handle that case?