OpenAnonymousBuffer: Wait for EOF before returning buffer.
This ensures that hooks/buffer-reload.cc will have fully run before the buffer
is returned. If we don't do this, //src:navigate_command has trouble with some
symbols in the VM environment that are defined yet not propertly initialized
yet.

I guess a more robust fix is to always initialize symbols after the compilation
phase is done, before execution. Some could be initialized directly to
hard-coded values, but in the general sense ... it's tricky, we may need to
execute code to get their initial values. So possibly one approach would be to
have an explicit "uninitialized" place holder.