Adjust BufferContentsViewLayout to better handle line-end.
Ranges that correspond to the end of the line now get extended to end at
std::numeric_limits<ColumnNumber>::max(). That means that any line-marks that
reference positions past the end of the line will be displayed in the metadata
for the last chunk for the line.

This happens often with C++ compiler emitting errors at the position right
at the end of the line.