OpenBuffer: Pass default_commands and mode directly to constructor.
This is strictly a no-op change.

Previously, they were build directly in the constructor. Now they are built in
the static factory method `OpenBuffer::New` and passed.

The reason for this change is that I expect to change the corresponding class
fields to gc::Ptr<> (rather than NonNull<std::shared_ptr<>>). By building them
outside of OpenBuffer, I know they will be retained in gc::Root pointers until
the point when a gc::Root<OpenBuffer> (that references them in its expansion)
has been built. Otherwise, if the gc::Root<> pointers are created in the
constructor (and only retained as gc::Ptr<> in the class fields), there's a race
condition where they could be collected before the buffer's gc::Root<> has been
created.