Avoid flickering of the help buffer.
Previously, each key-press (while the help buffer was/remained shown) would
cause a bit of flickering. This was due to the fact that a new anonymous buffer
had to be created, which requires async operations.

(Why does getting an anonymous buffer require async operations? Because we have
to execute buffer-reload.cc in the newly created buffer...)