Fix long-standing occasional crash on exit.
Turns out it was triggered by having multiple entries in `EditorState::buffers_`
that have different key (obviously, since `buffers_` is a map, you can't have
multiple entires with the same key) but the same value. When we computed the set
of values (for `data->pending_buffers`), we got a smaller set (since the value
was repeated). That meant that as we iterated `buffers_`, deleting from
`data->pending_buffers`, the 2nd time we got to the same value, we'd crash: the
buffer is already removed from `data->pending_buffers`.

Turns out this was caused by the logic around kFuturePasteBuffer, which was
inserting the buffer also under the name `BufferName::PasteBuffer()` (and
leaving it under the old name), thus having the same value under two different
keys.