EditorForTests: receive the config_path value (optionally).
This lets us strengthen our tests. In fact, it uncovered a bug with a race in
help_command.cc (and likely other places) with variables being declared but not
yet given a value (because the expression that gives them a value has yet to
run). We'll need to improve the handling for variables in that "limbo" state.