BufferRegistry::FindBuffersPathEndingIn: Implement.
This is a no-op, just moving logic from module file_link_mode to module
buffer_registry, but it will allow me to optimize an N^2 complexity operation in
a follow up (at the cost of keeping an additional index).

This N^2 complexity loop is taking significant time during the execution of
md::ExtractTags (which, in my Zettelkasten, is loading 5100+ buffers).