Integrate ExecutionContext::CompilationResult with gc.
This is incomplete. CompilationResult contains a pointer to a vm::Expression.

But at least it no longer contains a Root<Environment>.

We had cycles like this:

OpenBuffer -> MapMode -> CppCommand -> CompilationResult -> Environment
       ^                                                     ╷
       ╰─────────────────────────────────────────────────────╯

We still have loops, but I think we have fewer loops. I think we still have
loops through CompilationResult's dependency on vm::Expression.

To fix that we'll have to integrate Expression with gc, which... ugh.